Beginning the job of rebuilding the custom animations
Created a new animation named "weed_plant" that's intended to replace
the hacked together "weed_tree" anim that was modeled on the banana
tree. It still needs "grow", "chop", "burnt" and "chop_burnt"
animations, but these states are commented out of the prefab for now to
avoid errors.

Also, the pipe animation has a ton of problems so that's next on the
list for rebuild.