Solar Dryer is working - phase 1
It actually dries, but the mechanism is still a little funky. Since I'm
not using the built in Harvest mechanism, I'm trying to replace the
ingredients with their dried counterpart. Still having trying with
GiveItem to put things in the chest without an owner. For now, I'm
setting the slot to the item but that seems to bypass all kinds of
events and leads to bugs.