Solar Drying is working - phase 2
I figured out why it wasn't letting me do GiveItem - the dried_weed
prefab did not meet the requirements for the container. I've come up
with a clunky solution that I might streamline someday with tags, but
for now it works great. The final phase is to create some kind of "fruit
leather" or something for fruits/veggies to be dried into.